smart_platform_bigdata_iot_ai/requirements.txt
kafka-python
pyspark
tensorflow
keras
numpy